// frontend/src/app/interfaces/plan.ts
export interface Plan {
    headingColor: string;
    selected: boolean;
    planName: string;
    description: string;
    monthlyPrice: string;
    quality: string;
    resolutions: string;
    spatialAudio: string;
    supportedDevices: string;
    maxDeviceCount: number;
    downloadDeviceCount: number;
    ads: string;
  }
  